# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Machanic Shop API"
  description: "API to manage a mechanic shop."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: 
...
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: Mechanic login (generate token)
      description: Post login credentials (email and password), if valid a user specific token will be minted and returned.
      parameters:
        - in: body
          name: body
          description: Login Credentials
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        201:
          description: Successful login
          schema:
            type: object
            properties:
              token:
                type: string
          examples:
            application/json:
              message: "Successfully Logged In"
              status: "success"
              token: "eyJhbGciOiJI...."
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: Create a Mechanic
      description: Send a JSON payload with mechanic information to create a mechanic in the database.
      parameters:
        - in: body
          name: body
          description: Information to be created
          required: true
          schema:
            $ref: '#/definitions/MechanicPayload'
      responses:
        201:
          description: Successful creation with hashed password
          schema:
            $ref: '#/definitions/MechanicResponse'
          examples:
            application/json:
              email: "johndoe@email.com"
              id: 1
              name: "John Doe"
              password: "<hashed password>"
              phone: "123-123-1234"
              salary: 50000

    put:
      tags:
        - Mechanics
      summary: Update a Mechanic (Token Required)
      description: Update an existing Mechanic, requires that mechanic to be logged in.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: Information used to update mechanic
          required: true
          schema:
            $ref: '#/definitions/MechanicPayload'
      responses:
        201:
          description: Successful update mechanic with hashed password
          schema:
            $ref: '#/definitions/MechanicResponse'
          examples:
            application/json:
              email: "johndoe@email.com"
              id: 1
              name: "John Doe"
              password: "<hashed password>"
              phone: "123-123-1234"
              salary: 50000

    get:
      tags:
        - Mechanics
      summary: Get Mechanics
      description: Endpoint to return all mechanics
      responses:
        200:
          description: Successful retrieval
          schema:
            $ref: '#/definitions/MechanicListResponse'
          examples:
            application/json:
              [{
                email: "johndoe@email.com",
                id: 1,
                name: "John Doe",
                phone: "123-123-1111",
                salary: 50000
              },{
                email: "janedoe@email.com",
                id: 2,
                name: "Jane Doe",
                phone: "123-123-2222",
                salary: 50000
              }]

    delete:
      tags:
        - Mechanics
      summary: Delete a Mechanic (Token Required)
      description: Delete an existing Mechanic, requires that mechanic to be logged in.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successful deletion of Mechanic
          examples:
            application/json:
              message: "Mechanic <mechanic.name> was deleted."
  /mechanics/{id}:
    get:
      tags:
        - Mechanics
      summary: Get a specific Mechanic using ID
      description: Pass in a Mechanic ID as a path parameter and returns specific Mechanic
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The Mechanic ID
      responses: 
        200:
          description: Successful query
          schema:
            $ref: '#/definitions/MechanicResponse'
          examples:
            application/json:
              email: "johndoe@email.com"
              id: 1
              name: "John Doe"
              password: "<hashed password>"
              phone: "123-123-1234"
              salary: 50000 
  /mechanics/search:
    get:
      tags:
        - Mechanics
      summary: Get Mechanic(s) using search query by name
      description: Receives name as a query parameter and returns returns all Mechanics who's name matches the query parameter passed in.
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Name of Mechanic(s) to search.
      responses: 
        200:
          description: Successful search
          schema:
            $ref: '#/definitions/SearchMechanicList'
          examples:
            application/json:
              mechanics: [{
                email: "johndoe@email.com",
                id: 1,
                name: "John Doe",
                phone: "123-123-1111",
                salary: 50000
              },{
                email: "janedoe@email.com",
                id: 2,
                name: "Jane Doe",
                phone: "123-123-2222",
                salary: 50000
              }]
  /mechanics/paginated:
    get:
      tags:
        - Mechanics
      summary: Get Mechanic(s) using paginated query parameters
      description: Receives page and per_page as a query parameters and returns the number of Mechanic(s) as specified by per_page query parameters on the specified page query parameter.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: The page number to display Mechanic(s).
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
          description: The number of Mechanics to display per page.
      responses:
        200:
          description: Successful retrieval
          schema:
            $ref: '#/definitions/MechanicListResponse'
          examples:
            application/json:
              [{
                email: "johndoe@email.com",
                id: 1,
                name: "John Doe",
                phone: "123-123-1111",
                salary: 50000
              },{
                email: "janedoe@email.com",
                id: 2,
                name: "Jane Doe",
                phone: "123-123-2222",
                salary: 50000
              }]
...
definitions: 
...
  MechanicPayload:
    type: object
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      salary:
        type: integer
  MechanicResponse:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      salary:
        type: integer
  MechanicListResponse: #password excluded
    type: array
    items:
      type: object
      properties:
        email:
          type: string
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        salary:
          type: integer
  SearchMechanicList:
      type: object
      properties:
        mechanics:
          type: array
          items:
            type: object
            properties:
              email:
                type: string
              id:
                type: integer
              name:
                type: string
              phone:
                type: string
              salary:
                type: integer
...